c++: fix remaining MSVC compiler warnings

+++ core/src/Range.h
@@ -20,7 +20,7 @@ struct StrideIter
 	StrideIter<Iterator>& operator++() { return pos += stride, *this; }
 	bool operator<(const StrideIter<Iterator>& rhs) const { return pos < rhs.pos; }
 	StrideIter<Iterator> operator+(int i) const { return {pos + i * stride, stride}; }
-	int operator-(const StrideIter<Iterator>& rhs) const { return (pos - rhs.pos) / stride; }
+	int operator-(const StrideIter<Iterator>& rhs) const { return narrow_cast<int>((pos - rhs.pos) / stride); }
 };
 
 template <typename Iterator>
